Move grid to utilities
