+++ src/nationalarchives/components/_presentation.scss
@@ -9,7 +9,6 @@
 @use "footer";
 @use "gallery";
 @use "global-header";
-@use "grid";
 @use "header";
 @use "hero";
 @use "index-grid";
