--- src/nationalarchives/components/grid/_index.scss
@@ -1,1 +1,0 @@
-@use "grid";
