--- src/nationalarchives/components/grid/macro.njk
@@ -1,3 +1,0 @@
-{% macro tnaGrid(params) %}
-  {%- include "nationalarchives/components/grid/template.njk" -%}
-{% endmacro %}
