+++ src/nationalarchives/utilities/colour/colour-themes.stories.js
@@ -1,26 +1,26 @@
-import Accordion from "../../../components/accordion/template.njk";
-import Breadcrumbs from "../../../components/breadcrumbs/template.njk";
-import Button from "../../../components/button/template.njk";
-import Card from "../../../components/card/template.njk";
-import Checkboxes from "../../../components/checkboxes/template.njk";
-import ErrorSummary from "../../../components/error-summary/template.njk";
-import Footer from "../../../components/footer/template.njk";
-import Gallery from "../../../components/gallery/template.njk";
-import GlobalHeader from "../../../components/global-header/template.njk";
-import Hero from "../../../components/hero/template.njk";
-import IndexGrid from "../../../components/index-grid/template.njk";
-import Pagination from "../../../components/pagination/template.njk";
-import PhaseBanner from "../../../components/phase-banner/template.njk";
-import Picture from "../../../components/picture/template.njk";
-import Radios from "../../../components/radios/template.njk";
-import RecordsList from "../../../components/records-list/template.njk";
-import SearchField from "../../../components/search-field/template.njk";
-import Select from "../../../components/select/template.njk";
-import SkipLink from "../../../components/skip-link/template.njk";
-import Tabs from "../../../components/tabs/template.njk";
-import TextInput from "../../../components/text-input/template.njk";
-import Textarea from "../../../components/textarea/template.njk";
-import Warning from "../../../components/warning/template.njk";
+import Accordion from "../../components/accordion/template.njk";
+import Breadcrumbs from "../../components/breadcrumbs/template.njk";
+import Button from "../../components/button/template.njk";
+import Card from "../../components/card/template.njk";
+import Checkboxes from "../../components/checkboxes/template.njk";
+import ErrorSummary from "../../components/error-summary/template.njk";
+import Footer from "../../components/footer/template.njk";
+import Gallery from "../../components/gallery/template.njk";
+import GlobalHeader from "../../components/global-header/template.njk";
+import Hero from "../../components/hero/template.njk";
+import IndexGrid from "../../components/index-grid/template.njk";
+import Pagination from "../../components/pagination/template.njk";
+import PhaseBanner from "../../components/phase-banner/template.njk";
+import Picture from "../../components/picture/template.njk";
+import Radios from "../../components/radios/template.njk";
+import RecordsList from "../../components/records-list/template.njk";
+import SearchField from "../../components/search-field/template.njk";
+import Select from "../../components/select/template.njk";
+import SkipLink from "../../components/skip-link/template.njk";
+import Tabs from "../../components/tabs/template.njk";
+import TextInput from "../../components/text-input/template.njk";
+import Textarea from "../../components/textarea/template.njk";
+import Warning from "../../components/warning/template.njk";
 
 const argTypes = {
   theme: {
