+++ src/nationalarchives/utilities/forms/_index.scss
@@ -1,8 +1,8 @@
-@use "../utilities/typography" as typographyUtilities;
-@use "../tools/colour";
-@use "../tools/media";
-@use "../tools/spacing";
-@use "../tools/typography";
+@use "../../utilities/typography" as typographyUtilities;
+@use "../../tools/colour";
+@use "../../tools/media";
+@use "../../tools/spacing";
+@use "../../tools/typography";
 
 button,
 input,
