+++ src/nationalarchives/utilities/forms/forms.mdx
@@ -1,14 +1,14 @@
 import { Meta, Canvas } from "@storybook/blocks";
-import * as TextInputStories from "../../../components/text-input/text-input.stories";
-import * as TextareaStories from "../../../components/textarea/textarea.stories";
-import * as DateInputStories from "../../../components/date-input/date-input.stories";
-import * as DateSearchStories from "../../../components/date-search/date-search.stories";
-import * as SelectStories from "../../../components/select/select.stories";
-import * as RadiosStories from "../../../components/radios/radios.stories";
-import * as CheckboxesStories from "../../../components/checkboxes/checkboxes.stories";
-import * as SearchFieldStories from "../../../components/search-field/search-field.stories";
+import * as TextInputStories from "../../components/text-input/text-input.stories";
+import * as TextareaStories from "../../components/textarea/textarea.stories";
+import * as DateInputStories from "../../components/date-input/date-input.stories";
+import * as DateSearchStories from "../../components/date-search/date-search.stories";
+import * as SelectStories from "../../components/select/select.stories";
+import * as RadiosStories from "../../components/radios/radios.stories";
+import * as CheckboxesStories from "../../components/checkboxes/checkboxes.stories";
+import * as SearchFieldStories from "../../components/search-field/search-field.stories";
 
-<Meta title="Utilities/Form elements" />
+<Meta title="Utilities/Forms" />
 
 # Form elements
