+++ src/nationalarchives/utilities/grid/macro.njk
@@ -1,0 +1,3 @@
+{% macro tnaGrid(params) %}
+  {%- include "nationalarchives/utilities/grid/template.njk" -%}
+{% endmacro %}
