+++ src/nationalarchives/utilities/lists/_index.scss
@@ -1,8 +1,8 @@
-@use "../tools/colour";
-@use "../tools/media";
-@use "../tools/spacing";
-@use "../tools/typography";
-@use "../utilities/typography" as typographyUtils;
+@use "../../tools/colour";
+@use "../../tools/media";
+@use "../../tools/spacing";
+@use "../../tools/typography";
+@use "../../utilities/typography" as typographyUtils;
 
 .tna-ul {
   list-style-type: disc;
