+++ src/nationalarchives/utilities/tables/_index.scss
@@ -1,10 +1,10 @@
 @use "sass:math";
 @use "sass:selector";
-@use "../tools/colour";
-@use "../tools/grid";
-@use "../tools/media";
-@use "../tools/spacing";
-@use "../tools/typography";
+@use "../../tools/colour";
+@use "../../tools/grid";
+@use "../../tools/media";
+@use "../../tools/spacing";
+@use "../../tools/typography";
 
 .tna-table {
   width: 100%;
