+++ tasks/test-package.js
@@ -94,7 +94,6 @@ const checkExists = [
   ...componentFiles("footer", "Footer"),
   ...componentFiles("gallery", "Gallery"),
   ...componentFiles("global-header", "GlobalHeader"),
-  ...componentFiles("grid"),
   ...componentFiles("header", "Header"),
   ...componentFiles("hero"),
   ...componentFiles("index-grid"),
@@ -123,17 +122,23 @@ const checkExists = [
   // Utilities
   "nationalarchives/utilities/_a11y.scss",
   "nationalarchives/utilities/_areas.scss",
-  "nationalarchives/utilities/_colour.scss",
+  "nationalarchives/utilities/colour/_index.scss",
   "nationalarchives/utilities/_columns.scss",
   "nationalarchives/utilities/_debug.scss",
-  "nationalarchives/utilities/_forms.scss",
+  "nationalarchives/utilities/forms/_index.scss",
   "nationalarchives/utilities/_global.scss",
+  "nationalarchives/utilities/grid/_index.scss",
+  "nationalarchives/utilities/grid/fixtures.json",
+  "nationalarchives/utilities/grid/grid.stories.js",
+  "nationalarchives/utilities/grid/macro-options.json",
+  "nationalarchives/utilities/grid/macro.njk",
+  "nationalarchives/utilities/grid/template.njk",
   "nationalarchives/utilities/_index.scss",
-  "nationalarchives/utilities/_lists.scss",
-  "nationalarchives/utilities/_overrides.scss",
+  "nationalarchives/utilities/lists/_index.scss",
+  "nationalarchives/utilities/overrides/_index.scss",
   "nationalarchives/utilities/_reset.scss",
-  "nationalarchives/utilities/_tables.scss",
-  "nationalarchives/utilities/_typography.scss",
+  "nationalarchives/utilities/tables/_index.scss",
+  "nationalarchives/utilities/typography/_index.scss",
   // Variables
   "nationalarchives/variables/_a11y.scss",
   "nationalarchives/variables/_assets.scss",
